remove version specification for jack-sys in jack's Cargo.toml
jack 0.9.1 was published depending on jack-sys 0.3.3, which breaks
using the RUST_JACK_DLOPEN environment variable to activate the
dlopen feature. To fix this and avoid this error prone release
process, remove the version requirement in jack's Cargo.toml and
only use the relative path to jack-sys so they are always published
together.